requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
schedule==1.2.0
psutil==5.9.6
aiohttp==3.9.1
cachetools==5.3.2

